pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iowarp-hooks"
version = "0.1.0"
description = "Claude Code hook manager for easy installation and management of hook sets"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "rich>=12.0.0",
]

[project.scripts]
iowarp-hooks = "iowarp_hooks.hook_manager:main"

[project.urls]
Homepage = "https://github.com/yourusername/iowarp-hooks"
Repository = "https://github.com/yourusername/iowarp-hooks"

[tool.hatch.version]
path = "src/iowarp_hooks/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/iowarp_hooks"]